Split maflib.core into some smaller modules
