+++ sample_wscript
@@ -1,5 +1,7 @@
 import maf
-import maflib.core
+import maflib.plot
+import maflib.rules
+import maflib.util
 import subprocess
 
 def options(opt):
@@ -48,7 +50,7 @@ def experiment(exp):
 
     exp(source='news20.scale',
         target='model',
-        parameters=maflib.core.product({
+        parameters=maflib.util.product({
             's': [0, 1, 2, 3, 4, 7],
             'C': [0.001, 0.01, 0.1, 1, 10, 100],
             'B': [-1, 1]
@@ -61,18 +63,18 @@ def experiment(exp):
 
     exp(source='raw_result',
         target='result',
-        rule=maflib.core.convert_libsvm_accuracy)
+        rule=maflib.rules.convert_libsvm_accuracy)
 
     exp(source='result',
         target='max_accuracy',
         for_each=['s', 'C'],
         # aggregate_by=['B'],
-        rule=maflib.core.max('accuracy'))
+        rule=maflib.rules.max('accuracy'))
 
     exp(source='max_accuracy',
         target='figure.png',
         for_each=[],
-        rule=maflib.core.plot_line(
+        rule=maflib.plot.plot_line(
             x={'key': 'C', 'scale': 'log'},
             y='accuracy',
             legend={'key': 's', 'labels': LIBLINEAR_METHODS}))
@@ -80,12 +82,12 @@ def experiment(exp):
     # cv
     exp(source='news20.scale',
         target='cv_train cv_test',
-        parameters=maflib.core.product({'fold':range(5)}), # 5-fold cross validation
-        rule=maflib.core.segment_by_line(5)) # assume that a line corresponds to a example
+        parameters=maflib.util.product({'fold':range(5)}), # 5-fold cross validation
+        rule=maflib.rules.segment_by_line(5)) # assume that a line corresponds to a example
 
     exp(source='cv_train cv_test',
         target='cv_model cv_raw_result',
-        parameters=maflib.core.product({
+        parameters=maflib.util.product({
             's': [0, 1],
             'C': [0.01],
             'B': [1]
@@ -93,8 +95,8 @@ def experiment(exp):
         rule=liblinear_train_test)
     exp(source='cv_raw_result',
         target='cv_result',
-        rule=maflib.core.convert_libsvm_accuracy)
+        rule=maflib.rules.convert_libsvm_accuracy)
     exp(source='cv_result',
         target='cv_ave_accuracy',
         for_each=['s', 'C', 'B'],
-        rule=maflib.core.average())
+        rule=maflib.rules.average())
